refactor(games): extract triathlon, capital, and cricket models
Move dense game logic out of the triathlon utility and the Capital/Cricket views into dedicated domain and feature modules while preserving runtime behavior.

Changes included:
- extract triathlon scorecard types and scoring rules into src/domain/triathlon
- keep utils/triathlonScoring as a compatible facade over the new domain modules
- extract Capital reducer, history snapshots, and result helpers into capitalGameModel
- extract Cricket competitor, snapshot, aggregate, summary, and turn helpers into cricketGameModel
- reduce CapitalGameView and CricketGameView to rendering and screen-level handlers
- add focused unit tests for triathlon, Capital, and Cricket model slices

Validation already completed:
- npx vitest run tests/unit/triathlonScoring.test.ts
- npx vitest run tests/unit/capitalGameModel.test.ts tests/unit/cricketGameModel.test.ts
- npm run test:unit
- npm run typecheck
