+++ src/domain/triathlon/index.ts
@@ -1,0 +1,2 @@
+export * from './triathlonTypes';
+export * from './triathlonScoringRules';
